+++ source/Annex.Core/Input/InputEvents/MouseButtonReleasedEvent.cs
@@ -2,10 +2,10 @@
 {
     public class MouseButtonReleasedEvent : MouseButtonEvent
     {
-        public int WindowX { get; }
-        public int WindowY { get; }
+        public float WindowX { get; }
+        public float WindowY { get; }
 
-        public MouseButtonReleasedEvent(MouseButton button, int x, int y) : base(button) {
+        public MouseButtonReleasedEvent(MouseButton button, float x, float y) : base(button) {
             this.WindowX = x;
             this.WindowY = y;
         }
